+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # DEV
+- Use setup.cfg until Flit supports `extras_require`
 - Add default registration serializer with casefold
 # 0.1.0
 - Remove revproxy from defaults